requirements.txt
pulp
mysql-connector-python